Enable mock pcrf for relavent tests
