+++ lte/gateway/python/integ_tests/s1aptests/test_attach_detach_after_ue_context_release.py
@@ -10,7 +10,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
+import os
 import time
 import unittest
 
@@ -22,7 +22,8 @@ import s1ap_wrapper
 class TestAttachDetachAfterUeContextRelease(unittest.TestCase):
 
     def setUp(self):
-        self._s1ap_wrapper = s1ap_wrapper.TestWrapper()
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
+        self._s1ap_wrapper = s1ap_wrapper.TestWrapper(mock_pcrf=self.mock_pcrf)
 
     def tearDown(self):
         self._s1ap_wrapper.cleanup()
