+++ lte/gateway/python/integ_tests/s1aptests/test_attach_detach_dedicated_bearer_deactivation_invalid_imsi.py
@@ -1,6 +1,5 @@
 """
 Copyright 2020 The Magma Authors.
-
 This source code is licensed under the BSD-style license found in the
 LICENSE file in the root directory of this source tree.
 
@@ -10,8 +9,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
 import ipaddress
+import os
 import time
 import unittest
 
@@ -27,7 +26,8 @@ class TestAttachDetachDedicatedBearerDeactivationInvalidImsi(unittest.TestCase):
 
     def setUp(self):
         """Initialize"""
-        self._s1ap_wrapper = s1ap_wrapper.TestWrapper()
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
+        self._s1ap_wrapper = s1ap_wrapper.TestWrapper(mock_pcrf=self.mock_pcrf)
         self._spgw_util = SpgwUtil()
 
     def tearDown(self):
