+++ lte/gateway/python/integ_tests/s1aptests/test_attach_detach_multi_ue.py
@@ -11,6 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import os
 import unittest
 
 import s1ap_types
@@ -20,7 +21,8 @@ import s1ap_wrapper
 class TestAttachDetachMultiUe(unittest.TestCase):
 
     def setUp(self):
-        self._s1ap_wrapper = s1ap_wrapper.TestWrapper()
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
+        self._s1ap_wrapper = s1ap_wrapper.TestWrapper(mock_pcrf=self.mock_pcrf)
 
     def tearDown(self):
         self._s1ap_wrapper.cleanup()
