+++ lte/gateway/python/integ_tests/s1aptests/test_attach_detach_security_algo_eea2_eia2.py
@@ -10,7 +10,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
+import os
 import unittest
 
 import s1ap_types
@@ -22,7 +22,8 @@ class TestAttachDetachSecurityAlgoEea2Eia2(unittest.TestCase):
 
     def setUp(self):
         """Initialize before test case execution"""
-        self._s1ap_wrapper = s1ap_wrapper.TestWrapper()
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
+        self._s1ap_wrapper = s1ap_wrapper.TestWrapper(mock_pcrf=self.mock_pcrf)
 
     def tearDown(self):
         """Cleanup after test case execution"""
@@ -56,7 +57,7 @@ class TestAttachDetachSecurityAlgoEea2Eia2(unittest.TestCase):
             config_list[i].ueNwCap_pr.eia2_128 = 1
             config_list[i].ueNwCap_pr.eia1_128 = 0
             config_list[i].ueNwCap_pr.eia0 = 0
-        self._s1ap_wrapper.configUEDevice(num_ues, req_data=config_list)
+        self._s1ap_wrapper.configUEDevice(num_ues, req_data=config_list, mock_pcrf=self.mock_pcrf)
 
         for i in range(num_ues):
             req = self._s1ap_wrapper.ue_req
