+++ lte/gateway/python/integ_tests/s1aptests/test_attach_detach_with_mme_restart.py
@@ -10,7 +10,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
+import os
 import time
 import unittest
 
@@ -22,8 +22,10 @@ from s1ap_utils import MagmadUtil
 class TestAttachDetachWithMmeRestart(unittest.TestCase):
 
     def setUp(self):
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
         self._s1ap_wrapper = s1ap_wrapper.TestWrapper(
             stateless_mode=MagmadUtil.stateless_cmds.ENABLE,
+            mock_pcrf=self.mock_pcrf,
         )
 
     def tearDown(self):
