+++ lte/gateway/python/integ_tests/s1aptests/test_attach_ipv4v6_pdn_type.py
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-
+import os
 import unittest
 
 import s1ap_types
@@ -23,7 +23,8 @@ class TestAttachIpv4v6PdnType(unittest.TestCase):
 
     def setUp(self):
         """Initialize before test case execution"""
-        self._s1ap_wrapper = s1ap_wrapper.TestWrapper()
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
+        self._s1ap_wrapper = s1ap_wrapper.TestWrapper(mock_pcrf=self.mock_pcrf)
 
     def tearDown(self):
         """Cleanup after test case execution"""
