+++ lte/gateway/python/integ_tests/s1aptests/test_no_auth_response_with_mme_restart.py
@@ -11,6 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import os
 import time
 import unittest
 
@@ -24,8 +25,10 @@ from s1ap_utils import MagmadUtil
 
 class TestNoAuthResponseWithMmeRestart(unittest.TestCase):
     def setUp(self):
+        self.mock_pcrf = {True if (os.environ.get("FEDERATED_MODE") == "True") else False}
         self._s1ap_wrapper = s1ap_wrapper.TestWrapper(
             stateless_mode=MagmadUtil.stateless_cmds.ENABLE,
+            mock_pcrf=self.mock_pcrf,
         )
 
     def tearDown(self):
